+++ src/Stryker.Core/Stryker.Core.UnitTest/MutantFilters/BaselineMutantFilterTests.cs
@@ -13,7 +13,7 @@ using Stryker.Core.ProjectComponents.TestProjects;
 using Stryker.Core.Reporters;
 using Stryker.Core.Reporters.Json;
 using Stryker.Core.Reporters.Json.SourceFiles;
-using Stryker.Core.UnitTest.Reporters;
+using Stryker.Core.UnitTest.Reporters.Json;
 using Xunit;
 
 namespace Stryker.Core.UnitTest.MutantFilters