+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/Json/JsonReporterTests.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.IO.Abstractions;
 using System.IO.Abstractions.TestingHelpers;
 using System.Linq;
+using System.Runtime.InteropServices;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -19,7 +20,7 @@ using Stryker.Core.Reporters.Json;
 using Stryker.Core.Reporters.Json.SourceFiles;
 using Xunit;
 
-namespace Stryker.Core.UnitTest.Reporters
+namespace Stryker.Core.UnitTest.Reporters.Json
 {
     public class JsonReporterTests : TestBase
     {
@@ -39,10 +40,7 @@ namespace ExtraProject.XUnit
     }
 }
 ";
-        public JsonReporterTests()
-        {
-            _fileSystemMock.File.WriteAllText(_testFilePath, _testFileContents);
-        }
+        public JsonReporterTests() => _fileSystemMock.File.WriteAllText(_testFilePath, _testFileContents);
 
         [Fact]
         public void JsonMutantPositionLine_ThrowsArgumentExceptionWhenSetToLessThan1()
@@ -75,7 +73,7 @@ namespace ExtraProject.XUnit
         }
 
         [Fact]
-        public void JsonReportFileComponent_ShouldHaveLanguageSetToCs()
+        public void JsonReportFileComponent_ShouldHaveLanguageSet()
         {
             var folderComponent = ReportTestHelper.CreateProjectWith();
             var fileComponent = (CsharpFileLeaf)(folderComponent as CsharpFolderComposite).GetAllFiles().First();
@@ -98,7 +96,7 @@ namespace ExtraProject.XUnit
             var folderComponent = ReportTestHelper.CreateProjectWith();
             foreach (var file in (folderComponent as CsharpFolderComposite).GetAllFiles())
             {
-                var jsonReportComponent = new SourceFile(((CsharpFileLeaf)file));
+                var jsonReportComponent = new SourceFile((CsharpFileLeaf)file);
                 foreach (var mutant in file.Mutants)
                 {
                     jsonReportComponent.Mutants.ShouldContain(m => m.Id == mutant.Id.ToString());
@@ -113,7 +111,7 @@ namespace ExtraProject.XUnit
             var folderComponent = ReportTestHelper.CreateProjectWith(duplicateMutant: true);
             foreach (var file in (folderComponent as CsharpFolderComposite).GetAllFiles())
             {
-                var jsonReportComponent = new SourceFile(((CsharpFileLeaf)file), loggerMock);
+                var jsonReportComponent = new SourceFile((CsharpFileLeaf)file, loggerMock);
                 foreach (var mutant in file.Mutants)
                 {
                     jsonReportComponent.Mutants.ShouldContain(m => m.Id == mutant.Id.ToString());
@@ -151,6 +149,17 @@ namespace ExtraProject.XUnit
             var report = JsonReport.Build(new StrykerOptions(), folderComponent, It.IsAny<TestProjectsInfo>());
 
             report.ProjectRoot.ShouldBe("/home/user/src/project/");
+        }
+
+        [Fact]
+        public void JsonReport_ShouldContainFullPath()
+        {
+            var folderComponent = ReportTestHelper.CreateProjectWith(root: RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? "c://" : "/");
+
+            var report = JsonReport.Build(new StrykerOptions(), folderComponent, It.IsAny<TestProjectsInfo>());
+            var path = report.Files.Keys.First();
+
+            Path.IsPathFullyQualified(path).ShouldBeTrue($"{path} should not be a relative path");
         }
 
         [Fact]