+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/Json/MockJsonReport.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Stryker.Core.Reporters.Json;
 using Stryker.Core.Reporters.Json.SourceFiles;
 
-namespace Stryker.Core.UnitTest.Reporters
+namespace Stryker.Core.UnitTest.Reporters.Json
 {
     public class MockJsonReport : JsonReport
     {